run it, I dare you
made scrolling work better, and made a filtering system, and made it find the game.exe file with 100% accuracy, as long as there is one, few exceptions.
btw be careful running it, rn it will start every game on your pc, it crashed my monster pc, so do not run it.